Fix crashes: Graph.h uint32 shift UB, R=1 ST bucket overflow → hybrid bucket+set
Two root causes for segfaults at large s values:
1. Graph.h getEdgeCompressedId: (u << 32) where u is uint32_t → UB, key=0 on GCC
   Fix: static_cast<uint64_t>(u) << 32
2. R=1 ST: (int)countingV[i] overflows when support > 2.1B (e.g. C(109,6)=2.4B)
   Fix: hybrid bucket+overflow set with BUCKET_THRESHOLD=5M (same pattern as R≥3)

+++ src/graph/Graph.h
@@ -204,7 +204,7 @@ public:
         if (getNbrCount(u) < 1000) {
             return getEdgeIndex(u, v);
         }
-        const uint64_t key = (u << 32) | v;
+        const uint64_t key = (static_cast<uint64_t>(u) << 32) | static_cast<uint64_t>(v);
         const auto it = edgeIdMap_.find(key);
         if (it != edgeIdMap_.end()) return it->second;
         return static_cast<daf::Size>(-1);
@@ -419,7 +419,7 @@ public:
         if (getNbrCount(u) < 1000) {
             return getEdgeIndex(u, v);
         }
-        const uint64_t key = (u << 32) | v;
+        const uint64_t key = (static_cast<uint64_t>(u) << 32) | static_cast<uint64_t>(v);
         const auto it = edgeIdMap_.find(key);
         if (it != edgeIdMap_.end()) return it->second;
         return static_cast<daf::Size>(-1);
